feat(shared-directory): move public components to shared directory
