+++ apps/web/src/services/asset-inventory/collector/create-collector/modules/CreateCollectorStep3.vue
@@ -46,8 +46,8 @@ import {
 import DeleteModal from '@/common/components/modals/DeleteModal.vue';
 
 import AttachedServiceAccountForm
-    from '@/services/asset-inventory/collector/modules/AttachedServiceAccountForm.vue';
-import { useCollectorFormStore } from '@/services/asset-inventory/store/collector-form-store';
+    from '@/services/asset-inventory/collector/shared/AttachedServiceAccountForm.vue';
+import { useCollectorFormStore } from '@/services/asset-inventory/collector/shared/collector-form-store';
 
 
 const emit = defineEmits([
