+++ apps/web/src/services/asset-inventory/collector/shared/CollectorPluginContents.vue
