+++ apps/web/src/services/asset-inventory/collector/shared/CollectorScheduleForm.vue
